refactor(utils): unify resolver modules with functional design
主要改进：
- contextResolver 改为纯函数设计
  - ContextResolver 类 → resolveAppContext() 等函数
  - 新增 getUserId() 和 getProjectId()（从 tokenResolver 移入）
  - 统一的函数式 API，更易用

- tokenResolver 增强
  - 新增 getTokenStatus()（从 handlerHelpers 移入）
  - 新增 getTokenSourceLabel()
  - 新增 TokenSource enum
  - 统一 Token 相关的所有工具函数

- handlerHelpers 精简
  - 移除重复的 Token 状态函数（已移至 tokenResolver）
  - 保留参数合并相关函数（getEffectiveContext 等）
  - 通过 re-export 保持向后兼容

- server.ts 清理
  - 移除未使用的 import
  - 移除废弃的 setTransportMode
  - 使用 tokenResolver.hasToken() 检查认证

架构优化：
- ✅ 设计模式统一（都是纯函数）
- ✅ 职责清晰分层
  - contextResolver: Context 字段解析
  - tokenResolver: Token 加载和状态
  - handlerHelpers: 参数合并工具
- ✅ 减少代码重复（净减少 38 行）
- ✅ 易于测试和维护

向后兼容：
- handlerHelpers 通过 re-export 保持兼容
- 所有现有代码正常工作
